refactor: :recycle: refactor status bar integrations
