+++ packages/workspace-fusion/src/lib/components/index.ts
@@ -1,3 +1,2 @@
-export * from './StatusBarWrapper';
 export * from './CustomTabWrapper';
 export * from './Workspace';
