+++ packages/workspace-fusion/src/lib/integrations/status-bar/components/StatusBarWrapper.tsx
@@ -1,8 +1,8 @@
 import { StatusBar } from '@equinor/workspace-status-bar';
 
-import { FusionMediator } from '../types';
-import { StatusBarConfig } from '../integrations/status-bar';
-import { useFilteredData } from '../hooks/useFilteredData';
+import { FusionMediator } from '../../../types';
+import { StatusBarConfig } from '..';
+import { useFilteredData } from '../../../hooks/useFilteredData';
 
 type StatusBarWrapperProps<
 	TData extends Record<PropertyKey, unknown>,
