+++ packages/workspace-fusion/src/lib/integrations/status-bar/components/index.ts
@@ -1,0 +1,1 @@
+export * from './StatusBarWrapper';
