+++ packages/workspace-fusion/src/lib/integrations/status-bar/index.ts
@@ -1,4 +1,4 @@
-export * from './status-bar';
 export * from './types';
 export * from './utils';
 export * from './hooks';
+export * from './types/workspaceConfig';
