+++ packages/workspace-fusion/src/lib/integrations/status-bar/types/index.ts
@@ -1,1 +1,2 @@
 export * from './workspaceProps';
+export * from './workspaceConfig';
