+++ packages/workspace-fusion/src/lib/integrations/status-bar/types/workspaceConfig.ts
